Updated to be consistent with the common onboarding form released April 4th.
Minor changes to the 'goals' section to be easier and more explicit.
Moved the gps location and other information in the 'contact and address'
question set at the top, made some small changes to requires. Also deleted
the soil questions, as those results will be returned in more detail
elsewhere and on a field-by-field basis.